#151: Moved specific provider schemas closer to provider's packages & simplified naming for Glide schema
