+++ docs/swagger.json
@@ -95,7 +95,7 @@
                         "in": "body",
                         "required": true,
                         "schema": {
-                            "$ref": "#/definitions/schemas.UnifiedChatRequest"
+                            "$ref": "#/definitions/schemas.ChatRequest"
                         }
                     }
                 ],
@@ -103,7 +103,7 @@
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "$ref": "#/definitions/schemas.UnifiedChatResponse"
+                            "$ref": "#/definitions/schemas.ChatResponse"
                         }
                     },
                     "400": {
@@ -673,6 +673,52 @@
                 }
             }
         },
+        "schemas.ChatRequest": {
+            "type": "object",
+            "properties": {
+                "message": {
+                    "$ref": "#/definitions/schemas.ChatMessage"
+                },
+                "messageHistory": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/schemas.ChatMessage"
+                    }
+                },
+                "override": {
+                    "$ref": "#/definitions/schemas.OverrideChatRequest"
+                }
+            }
+        },
+        "schemas.ChatResponse": {
+            "type": "object",
+            "properties": {
+                "cached": {
+                    "type": "boolean"
+                },
+                "created": {
+                    "type": "integer"
+                },
+                "id": {
+                    "type": "string"
+                },
+                "model": {
+                    "type": "string"
+                },
+                "modelResponse": {
+                    "$ref": "#/definitions/schemas.ProviderResponse"
+                },
+                "model_id": {
+                    "type": "string"
+                },
+                "provider": {
+                    "type": "string"
+                },
+                "router": {
+                    "type": "string"
+                }
+            }
+        },
         "schemas.OverrideChatRequest": {
             "type": "object",
             "properties": {
@@ -712,52 +758,6 @@
                 },
                 "totalTokens": {
                     "type": "number"
-                }
-            }
-        },
-        "schemas.UnifiedChatRequest": {
-            "type": "object",
-            "properties": {
-                "message": {
-                    "$ref": "#/definitions/schemas.ChatMessage"
-                },
-                "messageHistory": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/schemas.ChatMessage"
-                    }
-                },
-                "override": {
-                    "$ref": "#/definitions/schemas.OverrideChatRequest"
-                }
-            }
-        },
-        "schemas.UnifiedChatResponse": {
-            "type": "object",
-            "properties": {
-                "cached": {
-                    "type": "boolean"
-                },
-                "created": {
-                    "type": "integer"
-                },
-                "id": {
-                    "type": "string"
-                },
-                "model": {
-                    "type": "string"
-                },
-                "modelResponse": {
-                    "$ref": "#/definitions/schemas.ProviderResponse"
-                },
-                "model_id": {
-                    "type": "string"
-                },
-                "provider": {
-                    "type": "string"
-                },
-                "router": {
-                    "type": "string"
                 }
             }
         }
