+++ pkg/providers/anthropic/client_test.go
@@ -56,7 +56,7 @@ func TestAnthropicClient_ChatRequest(t *testing.T) {
 	client, err := NewClient(providerCfg, clientCfg, telemetry.NewTelemetryMock())
 	require.NoError(t, err)
 
-	request := schemas.UnifiedChatRequest{Message: schemas.ChatMessage{
+	request := schemas.ChatRequest{Message: schemas.ChatMessage{
 		Role:    "human",
 		Content: "What's the biggest animal?",
 	}}
@@ -86,7 +86,7 @@ func TestAnthropicClient_BadChatRequest(t *testing.T) {
 	client, err := NewClient(providerCfg, clientCfg, telemetry.NewTelemetryMock())
 	require.NoError(t, err)
 
-	request := schemas.UnifiedChatRequest{Message: schemas.ChatMessage{
+	request := schemas.ChatRequest{Message: schemas.ChatMessage{
 		Role:    "human",
 		Content: "What's the biggest animal?",
 	}}
