+++ pkg/providers/bedrock/client_test.go
@@ -61,7 +61,7 @@ func TestBedrockClient_ChatRequest(t *testing.T) {
 	client, err := NewClient(providerCfg, clientCfg, telemetry.NewTelemetryMock())
 	require.NoError(t, err)
 
-	request := schemas.UnifiedChatRequest{Message: schemas.ChatMessage{
+	request := schemas.ChatRequest{Message: schemas.ChatMessage{
 		Role:    "user",
 		Content: "What's the biggest animal?",
 	}}
