+++ pkg/providers/bedrock/schemas.go
@@ -1,0 +1,11 @@
+package bedrock
+
+// Bedrock Chat Response
+type ChatCompletion struct {
+	InputTextTokenCount int `json:"inputTextTokenCount"`
+	Results             []struct {
+		TokenCount       int    `json:"tokenCount"`
+		OutputText       string `json:"outputText"`
+		CompletionReason string `json:"completionReason"`
+	} `json:"results"`
+}
