+++ pkg/providers/octoml/chat.go
@@ -9,6 +9,8 @@ import (
 	"net/http"
 	"time"
 
+	"glide/pkg/providers/openai"
+
 	"glide/pkg/providers/clients"
 
 	"glide/pkg/api/schemas"
@@ -47,7 +49,7 @@ func NewChatRequestFromConfig(cfg *Config) *ChatRequest {
 	}
 }
 
-func NewChatMessagesFromUnifiedRequest(request *schemas.UnifiedChatRequest) []ChatMessage {
+func NewChatMessagesFromUnifiedRequest(request *schemas.ChatRequest) []ChatMessage {
 	messages := make([]ChatMessage, 0, len(request.MessageHistory)+1)
 
 	// Add items from messageHistory first and the new chat message last
@@ -61,7 +63,7 @@ func NewChatMessagesFromUnifiedRequest(request *schemas.UnifiedChatRequest) []Ch
 }
 
 // Chat sends a chat request to the specified octoml model.
-func (c *Client) Chat(ctx context.Context, request *schemas.UnifiedChatRequest) (*schemas.UnifiedChatResponse, error) {
+func (c *Client) Chat(ctx context.Context, request *schemas.ChatRequest) (*schemas.ChatResponse, error) {
 	// Create a new chat request
 	chatRequest := c.createChatRequestSchema(request)
 
@@ -77,7 +79,7 @@ func (c *Client) Chat(ctx context.Context, request *schemas.UnifiedChatRequest) 
 	return chatResponse, nil
 }
 
-func (c *Client) createChatRequestSchema(request *schemas.UnifiedChatRequest) *ChatRequest {
+func (c *Client) createChatRequestSchema(request *schemas.ChatRequest) *ChatRequest {
 	// TODO: consider using objectpool to optimize memory allocation
 	chatRequest := c.chatRequestTemplate // hoping to get a copy of the template
 	chatRequest.Messages = NewChatMessagesFromUnifiedRequest(request)
@@ -85,7 +87,7 @@ func (c *Client) createChatRequestSchema(request *schemas.UnifiedChatRequest) *C
 	return chatRequest
 }
 
-func (c *Client) doChatRequest(ctx context.Context, payload *ChatRequest) (*schemas.UnifiedChatResponse, error) {
+func (c *Client) doChatRequest(ctx context.Context, payload *ChatRequest) (*schemas.ChatResponse, error) {
 	// Build request payload
 	rawPayload, err := json.Marshal(payload)
 	if err != nil {
@@ -152,7 +154,7 @@ func (c *Client) doChatRequest(ctx context.Context, payload *ChatRequest) (*sche
 	}
 
 	// Parse the response JSON
-	var openAICompletion schemas.OpenAIChatCompletion // Octo uses the same response schema as OpenAI
+	var openAICompletion openai.ChatCompletion // Octo uses the same response schema as OpenAI
 
 	err = json.Unmarshal(bodyBytes, &openAICompletion)
 	if err != nil {
@@ -161,7 +163,7 @@ func (c *Client) doChatRequest(ctx context.Context, payload *ChatRequest) (*sche
 	}
 
 	// Map response to UnifiedChatResponse schema
-	response := schemas.UnifiedChatResponse{
+	response := schemas.ChatResponse{
 		ID:       openAICompletion.ID,
 		Created:  openAICompletion.Created,
 		Provider: providerName,
