+++ pkg/providers/openai/client_test.go
@@ -56,7 +56,7 @@ func TestOpenAIClient_ChatRequest(t *testing.T) {
 	client, err := NewClient(providerCfg, clientCfg, telemetry.NewTelemetryMock())
 	require.NoError(t, err)
 
-	request := schemas.UnifiedChatRequest{Message: schemas.ChatMessage{
+	request := schemas.ChatRequest{Message: schemas.ChatMessage{
 		Role:    "user",
 		Content: "What's the biggest animal?",
 	}}
